basic security komid

+++ Server/src/main/java/is/hi/hbv501g/kosmosinn/Kosmosinn/RestControllers/BoardRestController.java
@@ -2,6 +2,7 @@ package is.hi.hbv501g.kosmosinn.Kosmosinn.RestControllers;
 
 import is.hi.hbv501g.kosmosinn.Kosmosinn.Entities.Board;
 import is.hi.hbv501g.kosmosinn.Kosmosinn.Entities.Topic;
+import is.hi.hbv501g.kosmosinn.Kosmosinn.Entities.User;
 import is.hi.hbv501g.kosmosinn.Kosmosinn.Services.BoardService;
 import is.hi.hbv501g.kosmosinn.Kosmosinn.Services.CommentService;
 import is.hi.hbv501g.kosmosinn.Kosmosinn.Services.TopicService;
@@ -10,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 import java.util.List;
 
@@ -28,6 +30,10 @@ public class BoardRestController {
         this.boardService = boardService;
         this.commentService = commentService;
     }
+
+    @Autowired
+    HttpSession session;
+
     @RequestMapping(value = "/api/boards", method = RequestMethod.GET)
     public List<Board> getAllBoards() {
         return boardService.findAll();
@@ -44,27 +50,36 @@ public class BoardRestController {
     }
     @PostMapping(value = "{id}/addBoard", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     public Board addBoard(@RequestBody Board board) {
-        boardService.save(board);
+        User currentUser = (User) session.getAttribute("loggedinuser");
+        if (userService.isAdmin(currentUser)) {
+            boardService.save(board);
+        }
         return board;
     }
 
     @PatchMapping(value = "{id}", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     public Board editBoard(@Valid @PathVariable("id") long id, @RequestBody Board editedBoard) {
+        User currentUser = (User) session.getAttribute("loggedinuser");
         Board board = boardService.findById(id).get();
-        if (editedBoard.getName() != null) {
-            board.setName(editedBoard.getName());
+        if (userService.isAdmin(currentUser)) {
+            if (editedBoard.getName() != null) {
+                board.setName(editedBoard.getName());
+            }
+            if (editedBoard.getDescription() != null) {
+                board.setDescription(editedBoard.getDescription());
+            }
+            boardService.save(board);
         }
-        if (editedBoard.getDescription() != null) {
-            board.setDescription(editedBoard.getDescription());
-        }
-        boardService.save(board);
         return board;
     }
 
     @DeleteMapping("{id}/delete")
     public void deleteBoard(@PathVariable("id") long id) {
-        Board board = boardService.findById(id).get();
-        boardService.delete(board);
+        User currentUser = (User) session.getAttribute("loggedinuser");
+        if (userService.isAdmin(currentUser)) {
+            Board board = boardService.findById(id).get();
+            boardService.delete(board);
+        }
     }
 
 }
